+++ src/main/java/com/ludo/study/studymatchingplatform/auth/kakao/service/dto/OauthUserSignupDto.java
@@ -1,6 +1,6 @@
 package com.ludo.study.studymatchingplatform.auth.kakao.service.dto;
 
-import com.ludo.study.studymatchingplatform.user.domain.Social;
+import com.ludo.study.studymatchingplatform.user.domain.user.Social;
 
 public record OauthUserSignupDto(
